Add implementation of fallback voice actions.
This commit adds a simple implementation of a fallback voice service,
which is used when a request to Assistant cannot be completed or fails.
For read actions, it reads out notification messages; for any other
action, it reads out an error message for.

The TTS package was adopted from the Car Messaging System App and
subsequently refactored.

Fix: 110587280
Test: manual
Change-Id: I2897e837a42154bcaab85f0b9a2b73185f599cfb