Fix bug in car-compat-message-notification criteria
REPLY is not required action for message notifications. Previously it was.
If REPLY ~is~ given in the notification, we check to see it has at least one
RemoteInput. However, when we removed the requirement, we didn't gate the
RemoteInput check based on the presence of REPLY. Therefore, even if a notification
meets our new requirements when not having REPLY, it will still fail this
classification, because the RemoteInput check will still require REPLY action.

Bug: 137654440
Test: manually verified that message notifications are classified correctly
 by having Messenger App send notifications without/with REPLY action.

Change-Id: I5679c1427b5b32e592ff23c9ba6f7bccdf78b45d