Fixed QC SeekBar clickable while disabled behavior
Because SeekBars don't respect drawable states for it's disabled
appearance, it needs to actually be disabled to appear correctly.
Because of this (and the fact that the AbsSeekBar parent class overrides
the View.java allowClickWhileDisabled behavior), it needs to be extended
to special case the clickable while disabled state when touched.

Bug: 205891123
Bug: 207175350
Test: manual, atest CarQCLibUnitTests
Change-Id: I3467334b0c3865dfd8872b6f1723da2eedca5316