Add support to load scalable ui panels from designcompose dcf file
Create a ScalableUiDoc class in the car-scalable-ui-designcompose-lib library that parses view content stored in designcompose files and creates data structures used to access data needed for scalable ui.

Create a PanelStateDocLoader class that uses the ScalableUiDoc class to load panels into a StateManager.

Add a test to test parsing of a DCF file into a list of PanelState objects.

Bug: 414813386
Test: atest CarScalableUIDesignComposeUnitTests
Flag: EXEMPT: reference code.
Change-Id: I579275560937762fbd9243bbf3f1e9f4885b3502