Add option to force transaction
During certain conflicting behavior, the state of the panel from the
ScalableUI perspective may be different than the WM perspective. By
forcing the transition to be within the transaction, the WM can get an
updated state even if ScalableUI doesn't believe it needs to provide
this notification.

Bug: 452380827
Test: manual
Flag: NONE bugfix
Change-Id: I33730389bb8fb521fb049f2fa608ec47057ffdb6